+++ packages/devtools_app/lib/src/performance/performance_controller.dart
@@ -4,8 +4,8 @@
 
 import 'dart:async';
 
+import '../profiler/cpu_profile_controller.dart';
 import '../profiler/cpu_profile_model.dart';
-import '../profiler/cpu_profiler_controller.dart';
 import '../ui/fake_flutter/fake_flutter.dart';
 import '../utils.dart';
