+++ packages/devtools_app/lib/src/profiler/cpu_profile_controller.dart
