+++ packages/devtools_app/lib/src/profiler/flutter/cpu_profiler.dart
@@ -4,8 +4,8 @@
 import 'package:flutter/material.dart';
 
 import '../../ui/fake_flutter/_real_flutter.dart';
+import '../cpu_profile_controller.dart';
 import '../cpu_profile_model.dart';
-import '../cpu_profiler_controller.dart';
 import 'cpu_profile_call_tree.dart';
 import 'cpu_profile_flame_chart.dart';
