+++ packages/devtools_app/lib/src/timeline/timeline_controller.dart
@@ -6,8 +6,8 @@ import 'dart:async';
 import '../auto_dispose.dart';
 import '../config_specific/logger/logger.dart';
 import '../globals.dart';
+import '../profiler/cpu_profile_controller.dart';
 import '../profiler/cpu_profile_transformer.dart';
-import '../profiler/cpu_profiler_controller.dart';
 import '../service_manager.dart';
 import '../ui/fake_flutter/fake_flutter.dart';
 import 'timeline_model.dart';
