+++ packages/devtools_app/test/cpu_profiler_controller_test.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 import 'package:devtools_app/src/globals.dart';
-import 'package:devtools_app/src/profiler/cpu_profiler_controller.dart';
+import 'package:devtools_app/src/profiler/cpu_profile_controller.dart';
 import 'package:devtools_app/src/service_manager.dart';
 import 'package:devtools_testing/support/cpu_profile_test_data.dart';
 import 'package:test/test.dart';
